Cleanup cloudfoundry deploy
We should not have a manifest since we want cloudfoundry to detect that we want to use the build in version. Currently the staticfile build pack is being use due to the fact that this repo contains a Staticfile.

It's preferable not to have a manifest if possible. In this case we don't need one because everything else already works and is less code.

I'm also adding some .cfignores that make pushing much faster due to less cruft bein pushed.